%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{as.character.GenericDataFile}
\alias{as.character.GenericDataFile}
\alias{GenericDataFile.as.character}
\alias{as.character,GenericDataFile-method}

\title{Returns a short string describing the file}

\description{
 Returns a short string describing the file.
}

\usage{
\method{as.character}{GenericDataFile}(x, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{GenericDataFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
