%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ColumnNamesInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{ColumnNamesInterface}
\docType{class}
\alias{ColumnNamesInterface}


 \title{The ColumnNamesInterface class interface}

 \description{
  Package:  R.filesets \cr
\bold{Class ColumnNamesInterface}\cr

\code{\link[R.oo]{Interface}}\cr
\code{~~|}\cr
\code{~~+--}\emph{\code{ColumnNamesInterface}}\cr

\bold{Directly known subclasses:}\cr
\emph{\link[R.filesets]{GenericTabularFile}}, \emph{\link[R.filesets]{TabularTextFile}}\cr

public abstract class \bold{ColumnNamesInterface}\cr
extends \link[R.oo]{Interface}\cr


 }

 \usage{ColumnNamesInterface(...)}

 \arguments{
   \item{...}{Not used.}
 }

 \section{Fields and Methods}{
  \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{clearColumnNamesTranslator} \tab  -\cr
 \tab \code{\link[R.filesets:getColumnNames.ColumnNamesInterface]{getColumnNames}} \tab Gets the column names.\cr
 \tab \code{\link[R.filesets:nbrOfColumns.ColumnNamesInterface]{nbrOfColumns}} \tab Gets the number of columns.\cr
 \tab \code{\link[R.filesets:setColumnNames.ColumnNamesInterface]{setColumnNames}} \tab Sets the column names.\cr
 \tab \code{setColumnNamesTranslator} \tab  -\cr
}


 \bold{Methods inherited from Interface}:\cr
extend, print, uses


 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{classes}
