%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.DEPRECATED.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Deprecated and defunct objects}
\alias{Deprecated and defunct objects}
\title{Deprecated and defunct objects}


\description{
 The following objects are \emph{defunct}:
 \itemize{
  \item None
 }

 The following objects are \emph{deprecated}:
 \itemize{
  \item None
 }
}


\keyword{documentation}
\keyword{internal}
