%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericTabularFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{dim.GenericTabularFile}
\alias{dim.GenericTabularFile}
\alias{GenericTabularFile.dim}
\alias{dim.GenericTabularFile}
\alias{dim,GenericTabularFile-method}


\title{Gets the dimension of data table}

\description{
 Gets the dimension of data table, which is the number of rows and the number of columns.
}

\usage{\method{dim}{GenericTabularFile}(x)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} of length two.
}
\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.filesets:nbrOfRows.GenericTabularFile]{*nbrOfRows}()}.
  \code{\link[R.filesets:nbrOfColumns.GenericTabularFile]{*nbrOfColumns}()}.
  For more information see \code{\link{GenericTabularFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
