%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getAlias.GenericDataFileSet}
\alias{getAlias.GenericDataFileSet}
\alias{GenericDataFileSet.getAlias}
\alias{getAlias.GenericDataFileSet}
\alias{getAlias,GenericDataFileSet-method}


\title{Gets the alias of the file set}

\description{
  Gets the alias of the file set.
}

\usage{\method{getAlias}{GenericDataFileSet}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}}, or \code{\link[base]{NULL}} if no alias is set.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.filesets:setAlias.GenericDataFileSet]{*setAlias}()}.
  \code{\link[R.filesets:getName.GenericDataFileSet]{*getName}()}.
  For more information see \code{\link{GenericDataFileSet}}.
}
\keyword{internal}
\keyword{methods}
