%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericTabularFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getColumnNames.GenericTabularFile}
\alias{getColumnNames.GenericTabularFile}
\alias{GenericTabularFile.getColumnNames}
\alias{getColumnNames.GenericTabularFile}
\alias{getColumnNames,GenericTabularFile-method}


\title{Gets the column names}

\description{
 Gets the column names, either by inferring the from the file or using the
 preset column names.
}

\usage{\method{getColumnNames}{GenericTabularFile}(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns \code{\link[base]{character}} \code{\link[base]{vector}}.
}
\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenericTabularFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
