%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getCreatedOn.GenericDataFile}
\alias{getCreatedOn.GenericDataFile}
\alias{GenericDataFile.getCreatedOn}
\alias{getCreatedOn.GenericDataFile}
\alias{getCreatedOn,GenericDataFile-method}


\title{Gets when the file was created}

\description{
  Gets when the file was created.
}

\usage{\method{getCreatedOn}{GenericDataFile}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{POSIXct} time stamp.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[base]{file.info}}().
  For more information see \code{\link{GenericDataFile}}.
}
\keyword{internal}
\keyword{methods}
