%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ColumnNamesInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getDefaultColumnNames.ColumnNamesInterface}
\alias{getDefaultColumnNames.ColumnNamesInterface}
\alias{ColumnNamesInterface.getDefaultColumnNames}
\alias{getDefaultColumnNames,ColumnNamesInterface-method}

\title{Gets the default column names}

\description{
  Gets the default column names, that is, the column names without translations.
}

\usage{
\method{getDefaultColumnNames}{ColumnNamesInterface}(...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.filesets:getColumnNames.ColumnNamesInterface]{*getColumnNames}()}.
  For more information see \code{\link{ColumnNamesInterface}}.
}
\keyword{internal}
\keyword{methods}
