%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getExtension.GenericDataFile}
\alias{getExtension.GenericDataFile}
\alias{GenericDataFile.getExtension}
\alias{getExtension,GenericDataFile-method}

\title{Gets the filename extension}

\description{
  Gets the filename extension.
}

\usage{
\method{getExtension}{GenericDataFile}(this, ...)
}

\arguments{
 \item{...}{Arguments passed to \code{\link[R.filesets:getFilename.GenericDataFile]{*getFilename}()} and
    \code{\link[R.filesets:getDefaultFullName.GenericDataFile]{*getDefaultFullName}()}.}
}

\value{
  Returns a \code{\link[base]{character}} string (which is of length zero if "empty" file).
}

\author{Henrik Bengtsson}

\seealso{
  For more information see \code{\link{GenericDataFile}}.
}
\keyword{internal}
\keyword{methods}
