%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getLastAccessedOn.GenericDataFile}
\alias{getLastAccessedOn.GenericDataFile}
\alias{GenericDataFile.getLastAccessedOn}
\alias{getLastAccessedOn.GenericDataFile}
\alias{getLastAccessedOn,GenericDataFile-method}


\title{Gets when the file was last accessed}

\description{
  Gets when the file was last accessed.
}

\usage{
\method{getLastAccessedOn}{GenericDataFile}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{POSIXct} time stamp.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[base]{file.info}}().
  For more information see \code{\link{GenericDataFile}}.
}
\keyword{internal}
\keyword{methods}
