%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getLastModifiedOn.GenericDataFile}
\alias{getLastModifiedOn.GenericDataFile}
\alias{GenericDataFile.getLastModifiedOn}
\alias{getLastModifiedOn,GenericDataFile-method}

\title{Gets when the file was last modified}

\description{
  Gets when the file was last modified.
}

\usage{
\method{getLastModifiedOn}{GenericDataFile}(this, ...)
}

\arguments{
 \item{...}{Not used.}
}

\value{
  Returns a \code{POSIXct} time stamp.
}

\author{Henrik Bengtsson}

\seealso{
  \code{\link[base]{file.info}}().
  For more information see \code{\link{GenericDataFile}}.
}
\keyword{internal}
\keyword{methods}
