%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getPath.GenericDataFile}
\alias{getPath.GenericDataFile}
\alias{GenericDataFile.getPath}
\alias{getPath.GenericDataFile}
\alias{getPath,GenericDataFile-method}


\title{Gets the path (directory) of the file}

\description{
  Gets the path (directory) of the file.
}

\usage{\method{getPath}{GenericDataFile}(this, ...)}

\arguments{
 \item{...}{Optional arguments passed to \code{\link[R.filesets:getPathname.GenericDataFile]{*getPathname}()}.}
}

\value{
  Returns a \code{\link[base]{character}} string (\code{\link[base]{NA}} if an "empty" file).
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenericDataFile}}.
}
\keyword{internal}
\keyword{methods}
