%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TabularTextFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasColumnHeader.TabularTextFile}
\alias{hasColumnHeader.TabularTextFile}
\alias{TabularTextFile.hasColumnHeader}
\alias{hasColumnHeader.TabularTextFile}
\alias{hasColumnHeader,TabularTextFile-method}


\title{Checks if there are column names in the header}

\description{
 Checks if there are column names in the header.
}

\usage{
\method{hasColumnHeader}{TabularTextFile}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{logical}}.
}
\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{TabularTextFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
