%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{length.GenericDataFileSet}
\alias{length.GenericDataFileSet}
\alias{GenericDataFileSet.length}
\alias{length.GenericDataFileSet}
\alias{length,GenericDataFileSet-method}

\alias{GenericDataFileSet.nbrOfFiles}
\alias{nbrOfFiles.GenericDataFileSet}
\alias{nbrOfFiles,GenericDataFileSet-method}


\title{Gets the number of files in the set}

\description{
  Gets the number of files in the set.
}

\usage{
 \method{length}{GenericDataFileSet}(x)
 \method{nbrOfFiles}{GenericDataFileSet}(this, ...)
}

\value{
  Returns an non-negative \code{\link[base]{integer}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenericDataFileSet}}.
}
\keyword{internal}
\keyword{methods}
