%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  ColumnNamesInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfColumns.ColumnNamesInterface}
\alias{nbrOfColumns.ColumnNamesInterface}
\alias{ColumnNamesInterface.nbrOfColumns}
\alias{nbrOfColumns,ColumnNamesInterface-method}

\title{Gets the number of columns}

\description{
 Gets the number of columns.
}

\usage{
\method{nbrOfColumns}{ColumnNamesInterface}(this, ...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}
\author{Henrik Bengtsson}

\seealso{
  \code{\link[R.filesets:getColumnNames.ColumnNamesInterface]{*getColumnNames}()}.
  For more information see \code{\link{ColumnNamesInterface}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
