%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericTabularFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfColumns.GenericTabularFile}
\alias{nbrOfColumns.GenericTabularFile}
\alias{GenericTabularFile.nbrOfColumns}
\alias{nbrOfColumns.GenericTabularFile}
\alias{nbrOfColumns,GenericTabularFile-method}


\title{Gets the number of columns}

\description{
 Gets the number of columns.
}

\usage{\method{nbrOfColumns}{GenericTabularFile}(this, ...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}
\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.filesets:nbrOfRows.GenericTabularFile]{*nbrOfRows}()}.
  \code{\link[R.filesets:dim.GenericTabularFile]{*dim}()}.
  For more information see \code{\link{GenericTabularFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
