%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericTabularFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{nbrOfRows.GenericTabularFile}
\alias{nbrOfRows.GenericTabularFile}
\alias{GenericTabularFile.nbrOfRows}
\alias{nbrOfRows,GenericTabularFile-method}

\title{Gets the number of data rows}

\description{
 Gets the number of data rows.
}

\usage{
\method{nbrOfRows}{GenericTabularFile}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns an \code{\link[base]{integer}}.
}
\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.filesets:dim.GenericTabularFile]{*dim}()}.
  For more information see \code{\link{GenericTabularFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
