%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericTabularFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readColumns.GenericTabularFile}
\alias{readColumns.GenericTabularFile}
\alias{GenericTabularFile.readColumns}
\alias{readColumns.GenericTabularFile}
\alias{readColumns,GenericTabularFile-method}


\title{Reads a subset of the columns as a data frame}

\description{
 Reads a subset of the columns as a data frame.
}

\usage{
\method{readColumns}{GenericTabularFile}(...)
}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{data.frame}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenericTabularFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
