%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericTabularFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readDataFrame.GenericTabularFile}
\alias{readDataFrame.GenericTabularFile}
\alias{GenericTabularFile.readDataFrame}
\alias{readDataFrame.GenericTabularFile}
\alias{readDataFrame,GenericTabularFile-method}


\title{Reads the tabular data as a data frame}

\description{
 Reads the tabular data as a data frame.
}

\usage{\method{readDataFrame}{GenericTabularFile}(...)}

\arguments{
  \item{...}{Not used.}
}

\value{
  Returns a \code{\link[base]{data.frame}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenericTabularFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
