%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  TabularTextFile.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{readLines.TabularTextFile}
\alias{readLines.TabularTextFile}
\alias{TabularTextFile.readLines}
\alias{readLines.TabularTextFile}
\alias{readLines,TabularTextFile-method}


\title{Reads the lines of the data file as strings}

\description{
 Reads the lines of the data file as strings.
}

\usage{\method{readLines}{TabularTextFile}(con, ...)}

\arguments{
  \item{...}{Optional arguments passed to \code{\link[base]{readLines}}().}
}

\value{
  Returns a \code{\link[base]{character}} \code{\link[base]{vector}}.
}
\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.filesets:readDataFrame.TabularTextFile]{*readDataFrame}()}.
  For more information see \code{\link{TabularTextFile}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
