%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{seq.GenericDataFileSet}
\alias{seq.GenericDataFileSet}
\alias{GenericDataFileSet.seq}
\alias{seq.GenericDataFileSet}
\alias{seq,GenericDataFileSet-method}


 \title{Gets an integer vector of file indices}

 \description{
   Gets an integer vector of file indices.
 }

 \usage{\method{seq}{GenericDataFileSet}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
   Returns an \code{\link[base]{integer}} \code{\link[base]{vector}} in [1,N] where N is the number of files,
   or an empty vector if the set is empty.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{GenericDataFileSet}}.
 }


\keyword{internal}
\keyword{methods}
