%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  FullNameInterface.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setTags.FullNameInterface}
\alias{setTags.FullNameInterface}
\alias{FullNameInterface.setTags}
\alias{setTags.FullNameInterface}
\alias{setTags,FullNameInterface-method}


\title{Sets the tags}

\description{
  Sets the tags.
}

\usage{
\method{setTags}{FullNameInterface}(this, tags="*", ...)
}

\arguments{
 \item{tags}{A \code{\link[base]{character}} \code{\link[base]{vector}} of tags.}
 \item{...}{Not used.}
}

\value{
  Returns (invisibly) itself.
}

\details{
  See \code{\link[R.filesets:getTags.FullNameInterface]{*getTags}()} for so called \emph{special tags}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.filesets:getTags.FullNameInterface]{*getTags}()}.
  For more information see \code{\link{FullNameInterface}}.
}
\keyword{internal}
\keyword{methods}
