%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setTags.GenericDataFileSet}
\alias{setTags.GenericDataFileSet}
\alias{GenericDataFileSet.setTags}
\alias{setTags.GenericDataFileSet}
\alias{setTags,GenericDataFileSet-method}


\title{Sets the tags of the file set}

\description{
  Sets the tags of the file set.
}

\usage{\method{setTags}{GenericDataFileSet}(this, tags="*", ...)}

\arguments{
 \item{tags}{A \code{\link[base]{character}} \code{\link[base]{vector}} of tags.}
 \item{...}{Not used.}
}

\value{
  Returns nothing.
}

\details{
  See \code{\link[R.filesets:getTags.GenericDataFileSet]{*getTags}()} for so called \emph{special tags}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  \code{\link[R.filesets:getTags.GenericDataFileSet]{*getTags}()}.
  For more information see \code{\link{GenericDataFileSet}}.
}
\keyword{internal}
\keyword{methods}
