%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.DEPRECATED.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Deprecated and defunct objects}
\alias{Deprecated and defunct objects}
\title{Deprecated and defunct objects}

\alias{getColumnNameTranslator.GenericTabularFile}
\alias{setColumnNameTranslator.GenericTabularFile}
\alias{getLabel.GenericDataFile}
\alias{setLabel.GenericDataFile}
\alias{getLabel.GenericDataFileSet}
\alias{setLabel.GenericDataFileSet}
\alias{getFileListV0.GenericDataFileSetList}
\alias{digest2}

\description{
 The following objects are \emph{defunct}:
 \itemize{
  \item getAlias(), setAlias()
  \item getLabel(), setLabel()
  \item getColumnNameTranslator(), setColumnNameTranslator()
  \item getFileListV0()
 }

 The following objects are \emph{deprecated}:
 \itemize{
  \item digest2()
 }
}


\keyword{documentation}
\keyword{internal}
