%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{extract.GenericDataFileSet}
\alias{extract.GenericDataFileSet}
\alias{GenericDataFileSet.extract}
\alias{extract.GenericDataFileSet}
\alias{extract,GenericDataFileSet-method}


\title{Extract a subset of the file set}

\description{
  Extract a subset of the file set.
}

\usage{\method{extract}{GenericDataFileSet}(this, files, ...)}

\arguments{
 \item{files}{An \code{\link[base]{integer}} or a \code{\link[base]{logical}} \code{\link[base]{vector}} indicating which data files
   to be extracted.}
 \item{...}{Not used.}
}

\value{
  Returns an \code{\link{GenericDataFileSet}} (or a subclass) object.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenericDataFileSet}}.
}
\keyword{internal}
\keyword{methods}
