%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  GenericDataFileSet.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getFile.GenericDataFileSet}
\alias{getFile.GenericDataFileSet}
\alias{GenericDataFileSet.getFile}
\alias{getFile.GenericDataFileSet}
\alias{getFile,GenericDataFileSet-method}


\title{Get a particular file of the file set}

\description{
 Get a particular file of the file set.
}

\usage{
\method{getFile}{GenericDataFileSet}(this, idx, ...)
}

\arguments{
  \item{idx}{An \code{\link[base]{integer}} index specifying the file to be returned.}
  \item{...}{Not used.}
}

\value{
  Returns a @GenericDataFile.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{GenericDataFileSet}}.
}



\keyword{internal}
\keyword{methods}
\keyword{IO}
\keyword{programming}
