%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{append}
\alias{append.default}
\alias{append.GenericDataFileSet}
\alias{appendFiles}
\alias{appendFiles.GenericDataFileSet}
\alias{as.character.GenericTabularFile}
\alias{as.character.TabularTextFile}
\alias{byName}
\alias{byName.GenericDataFileSet}
\alias{clearCache.GenericDataFileSet}
\alias{clone.GenericDataFile}
\alias{clone.GenericDataFileSet}
\alias{compareChecksum}
\alias{compareChecksum.GenericDataFile}
\alias{copyTo}
\alias{copyTo.GenericDataFile}
\alias{copyTo.GenericDataFileSet}
\alias{dim.GenericTabularFile}
\alias{dimension.TabularTextFile}
\alias{equals.GenericDataFile}
\alias{equals.GenericDataFileSet}
\alias{extractMatrix}
\alias{extractMatrix.GenericTabularFile}
\alias{extractMatrix.GenericTabularFileSet}
\alias{findByName}
\alias{findByName.GenericDataFileSet}
\alias{fromFile}
\alias{fromFile.GenericDataFile}
\alias{fromFiles}
\alias{GenericDataFileList}
\alias{GenericDataFileSetList}
\alias{GenericTabularFile}
\alias{getAlias}
\alias{getAlias.GenericDataFile}
\alias{getAttribute}
\alias{getAttribute.GenericDataFile}
\alias{getAttributes}
\alias{getAttributes.GenericDataFile}
\alias{getChecksum}
\alias{getChecksum.GenericDataFile}
\alias{getColumnNames}
\alias{getColumnNames.GenericTabularFile}
\alias{getColumnNames.TabularTextFile}
\alias{getColumnNameTranslator}
\alias{getColumnNameTranslator.GenericTabularFile}
\alias{getCreatedOn}
\alias{getCreatedOn.GenericDataFile}
\alias{getFile}
\alias{getFile.GenericDataFileSet}
\alias{getFileClass}
\alias{getFileClass.GenericDataFileSet}
\alias{getFileClass.GenericDataFileSetList}
\alias{getFileList}
\alias{getFileList.GenericDataFileSetList}
\alias{getFilename}
\alias{getFiles}
\alias{getFiles.GenericDataFileSet}
\alias{getFileSize}
\alias{getFileSize.GenericDataFile}
\alias{getFileSize.GenericDataFileSet}
\alias{getFileType}
\alias{getFullName}
\alias{getFullNames}
\alias{getFullNames.GenericDataFileSet}
\alias{getFullNameTranslator}
\alias{getFullNameTranslator.GenericDataFile}
\alias{getFullNameTranslator.GenericDataFileSet}
\alias{getHeader}
\alias{getHeader.TabularTextFile}
\alias{getLabel.GenericDataFile}
\alias{getLastAccessedOn}
\alias{getLastAccessedOn.GenericDataFile}
\alias{getLastModifiedOn}
\alias{getLastModifiedOn.GenericDataFile}
\alias{getNames}
\alias{getPathname}
\alias{getPathnames}
\alias{getReadArguments}
\alias{getReadArguments.TabularTextFile}
\alias{getTags}
\alias{gunzip.GenericDataFile}
\alias{gzip.GenericDataFile}
\alias{hasBeenModified}
\alias{hasBeenModified.GenericDataFile}
\alias{hasColumnHeader}
\alias{hasColumnHeader.TabularTextFile}
\alias{hasFile}
\alias{hasFile.GenericDataFileSet}
\alias{hasTag}
\alias{hasTag.GenericDataFile}
\alias{hasTag.GenericDataFileSet}
\alias{hasTags}
\alias{hasTags.GenericDataFile}
\alias{hasTags.GenericDataFileSet}
\alias{indexOf}
\alias{indexOf.GenericDataFileSet}
\alias{isFile.GenericDataFile}
\alias{nbrOfColumns}
\alias{nbrOfColumns.GenericTabularFile}
\alias{nbrOfFiles}
\alias{nbrOfLines}
\alias{nbrOfLines.TabularTextFile}
\alias{nbrOfRows}
\alias{nbrOfRows.GenericTabularFile}
\alias{nbrOfRows.TabularTextFile}
\alias{readChecksum}
\alias{readChecksum.GenericDataFile}
\alias{readColumnNames}
\alias{readColumnNames.TabularTextFile}
\alias{readColumns}
\alias{readColumns.GenericTabularFile}
\alias{readColumns.TabularTextFile}
\alias{readDataFrame}
\alias{readDataFrame.GenericTabularFile}
\alias{readDataFrame.TabularTextFile}
\alias{readLines}
\alias{readLines.default}
\alias{readLines.TabularTextFile}
\alias{readRawHeader}
\alias{readRawHeader.TabularTextFile}
\alias{readTableHeader}
\alias{readTableHeader.default}
\alias{renameTo}
\alias{renameTo.GenericDataFile}
\alias{renameToUpperCaseExt}
\alias{renameToUpperCaseExt.GenericDataFile}
\alias{reorder.GenericDataFileSet}
\alias{sapply}
\alias{sapply.default}
\alias{sapply.GenericDataFileSet}
\alias{setAlias}
\alias{setAlias.GenericDataFile}
\alias{setAttribute}
\alias{setAttribute.GenericDataFile}
\alias{setAttributes}
\alias{setAttributes.GenericDataFile}
\alias{setAttributesBy}
\alias{setAttributesBy.GenericDataFile}
\alias{setAttributesByTags}
\alias{setAttributesByTags.GenericDataFile}
\alias{setColumnNameTranslator}
\alias{setColumnNameTranslator.GenericTabularFile}
\alias{setFullName}
\alias{setFullName.GenericDataFile}
\alias{setFullName.GenericDataFileSet}
\alias{setFullNamesTranslator}
\alias{setFullNamesTranslator.GenericDataFileSet}
\alias{setFullNameTranslator}
\alias{setFullNameTranslator.GenericDataFile}
\alias{setFullNameTranslator.GenericDataFileSet}
\alias{setLabel.GenericDataFile}
\alias{setName}
\alias{setName.GenericDataFile}
\alias{setName.GenericDataFileSet}
\alias{setTags}
\alias{TabularTextFile}
\alias{testAttributes}
\alias{testAttributes.GenericDataFile}
\alias{translateColumnNames}
\alias{translateColumnNames.GenericTabularFile}
\alias{translateFullName}
\alias{translateFullName.GenericDataFile}
\alias{translateFullName.GenericDataFileSet}
\alias{update2}
\alias{update2.GenericDataFileSet}
\alias{validate.GenericDataFileSet}
\alias{validateChecksum}
\alias{validateChecksum.GenericDataFile}
\alias{verify}
\alias{verify.GenericTabularFile}
\alias{verify.TabularTextFile}
\alias{writeChecksum}
\alias{writeChecksum.GenericDataFile}
\alias{writeColumnsToFiles}
\alias{digest2}
\alias{getFilename.Arguments}
\alias{byPath}
\alias{getFullNameTranslator.FullNameInterface}
\alias{getDefaultFullName}
\alias{hasTag.FullNameInterface}
\alias{hasTags.FullNameInterface}
\alias{setFullName.FullNameInterface}
\alias{setFullNamesTranslatorByfunction}
\alias{setFullNamesTranslatorByfunction.GenericDataFileSet}
\alias{setFullNamesTranslatorByNULL}
\alias{setFullNamesTranslatorByNULL.GenericDataFileSet}
\alias{setFullNameTranslator.FullNameInterface}
\alias{setFullNameTranslatorByfunction}
\alias{setFullNameTranslatorByfunction.FullNameInterface}
\alias{setFullNameTranslatorByNULL}
\alias{setFullNameTranslatorByNULL.FullNameInterface}
\alias{setName.FullNameInterface}
\alias{translateFullName.FullNameInterface}
\alias{updateFullName}
\alias{updateFullName.FullNameInterface}
\alias{updateFullName.GenericDataFileSet}
\alias{updateFullNames}
\alias{updateFullNames.GenericDataFileSet}
\alias{[.TabularTextFile}
\alias{appendFullNamesTranslator}
\alias{appendFullNamesTranslator.GenericDataFileSet}
\alias{appendFullNamesTranslatorByfunction}
\alias{appendFullNamesTranslatorByfunction.GenericDataFileSet}
\alias{appendFullNamesTranslatorByNULL}
\alias{appendFullNamesTranslatorByNULL.GenericDataFileSet}
\alias{appendFullNameTranslator}
\alias{appendFullNameTranslator.FullNameInterface}
\alias{appendFullNameTranslatorBycharacter}
\alias{appendFullNameTranslatorBycharacter.FullNameInterface}
\alias{appendFullNameTranslatorByfunction}
\alias{appendFullNameTranslatorByfunction.FullNameInterface}
\alias{appendFullNameTranslatorByNULL}
\alias{appendFullNameTranslatorByNULL.FullNameInterface}
\alias{clearFullNamesTranslator}
\alias{clearFullNamesTranslator.GenericDataFileSet}
\alias{clearFullNameTranslator}
\alias{clearFullNameTranslator.FullNameInterface}
\alias{clearListOfFullNameTranslators}
\alias{clearListOfFullNameTranslators.FullNameInterface}
\alias{getListOfFullNameTranslators}
\alias{getListOfFullNameTranslators.FullNameInterface}
\alias{setListOfFullNameTranslators}
\alias{setListOfFullNameTranslators.FullNameInterface}
\alias{appendFullNameTranslatorBylist}
\alias{appendFullNameTranslatorBylist.FullNameInterface}
\alias{appendFullNamesTranslatorBylist}
\alias{appendFullNamesTranslatorBylist.GenericDataFileSet}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
