%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{GenericDataFileList}
\alias{[.TabularTextFile}
\alias{append}
\alias{append.default}
\alias{appendFiles}
\alias{appendFullNameTranslator}
\alias{appendFullNameTranslator.FullNameInterface}
\alias{appendFullNameTranslatorByNULL}
\alias{appendFullNameTranslatorByNULL.FullNameInterface}
\alias{appendFullNameTranslatorByTabularTextFile}
\alias{appendFullNameTranslatorByTabularTextFile.FullNameInterface}
\alias{appendFullNameTranslatorByTabularTextFileSet}
\alias{appendFullNameTranslatorByTabularTextFileSet.FullNameInterface}
\alias{appendFullNameTranslatorBycharacter}
\alias{appendFullNameTranslatorBycharacter.FullNameInterface}
\alias{appendFullNameTranslatorBydata.frame}
\alias{appendFullNameTranslatorBydata.frame.FullNameInterface}
\alias{appendFullNameTranslatorByfunction}
\alias{appendFullNameTranslatorByfunction.FullNameInterface}
\alias{appendFullNameTranslatorBylist}
\alias{appendFullNameTranslatorBylist.FullNameInterface}
\alias{appendFullNamesTranslator}
\alias{appendFullNamesTranslator.GenericDataFileSet}
\alias{appendFullNamesTranslatorByNULL}
\alias{appendFullNamesTranslatorByNULL.GenericDataFileSet}
\alias{appendFullNamesTranslatorByTabularTextFile}
\alias{appendFullNamesTranslatorByTabularTextFile.GenericDataFileSet}
\alias{appendFullNamesTranslatorByTabularTextFileSet}
\alias{appendFullNamesTranslatorByTabularTextFileSet.GenericDataFileSet}
\alias{appendFullNamesTranslatorBydata.frame}
\alias{appendFullNamesTranslatorBydata.frame.GenericDataFileSet}
\alias{appendFullNamesTranslatorByfunction}
\alias{appendFullNamesTranslatorByfunction.GenericDataFileSet}
\alias{appendFullNamesTranslatorBylist}
\alias{appendFullNamesTranslatorBylist.GenericDataFileSet}
\alias{as.GenericDataFileSetList}
\alias{as.GenericDataFileSetList.GenericDataFileSetList}
\alias{as.character.GenericDataFileSetList}
\alias{as.character.GenericTabularFile}
\alias{as.character.TabularTextFile}
\alias{as.data.frame.GenericDataFileSetList}
\alias{as.list.GenericDataFileSetList}
\alias{assertDuplicates}
\alias{assertDuplicates.GenericDataFileSetList}
\alias{byName}
\alias{byPath}
\alias{clearCache.GenericDataFileSet}
\alias{clearFullNameTranslator}
\alias{clearFullNameTranslator.FullNameInterface}
\alias{clearFullNamesTranslator}
\alias{clearFullNamesTranslator.GenericDataFileSet}
\alias{clearListOfFullNameTranslators}
\alias{clearListOfFullNameTranslators.FullNameInterface}
\alias{clone.GenericDataFile}
\alias{clone.GenericDataFileSet}
\alias{clone.GenericDataFileSetList}
\alias{compareChecksum}
\alias{copyTo}
\alias{digest2}
\alias{dimension.TabularTextFile}
\alias{dropRootPathTags}
\alias{dropRootPathTags.default}
\alias{dropTags}
\alias{dropTags.default}
\alias{equals.GenericDataFileSet}
\alias{extract.GenericDataFileSetList}
\alias{extractMatrix}
\alias{extractMatrix.GenericTabularFileSet}
\alias{findByName}
\alias{fromFile}
\alias{fromFiles}
\alias{fromFiles.GenericDataFileSet}
\alias{fullname}
\alias{fullname.default}
\alias{getAlias}
\alias{getAlias.GenericDataFile}
\alias{getAsteriskTags}
\alias{getAsteriskTags.GenericDataFileSetList}
\alias{getAttribute}
\alias{getAttribute.GenericDataFile}
\alias{getAttributes}
\alias{getAttributes.GenericDataFile}
\alias{getChecksum}
\alias{getChecksumObjects}
\alias{getColumnNameTranslator}
\alias{getColumnNameTranslator.GenericTabularFile}
\alias{getColumnNamesTranslator}
\alias{getColumnNamesTranslator.GenericTabularFile}
\alias{setColumnNamesTranslator}
\alias{setColumnNamesTranslator.GenericTabularFile}
\alias{setColumnNames}
\alias{setColumnNames.GenericTabularFile}
\alias{getColumnNames}
\alias{getCommentChar}
\alias{getCommentChar.TabularTextFile}
\alias{getCreatedOn}
\alias{getDefaultFullName}
\alias{getDefaultFullName.GenericDataFileSetList}
\alias{getDepth}
\alias{getDepth.GenericDataFileSet}
\alias{getExtension}
\alias{getExtensionPattern}
\alias{getExtensionPattern.GenericDataFile}
\alias{getFile}
\alias{getFileClass}
\alias{getFileClass.GenericDataFileSet}
\alias{getFileList}
\alias{getFileList.GenericDataFileSetList}
\alias{getFileListClass}
\alias{getFileListClass.GenericDataFileSetList}
\alias{getFileListV0}
\alias{getFileListV0.GenericDataFileSetList}
\alias{getFileSize}
\alias{getFileSize.GenericDataFileSet}
\alias{getFileType}
\alias{getFilenameExtension}
\alias{getFilenameExtension.GenericDataFile}
\alias{getFiles}
\alias{getFiles.GenericDataFileSet}
\alias{getFullName}
\alias{getFullNameTranslator}
\alias{getFullNameTranslator.FullNameInterface}
\alias{getFullNames}
\alias{getFullNames.GenericDataFileSetList}
\alias{getHeader}
\alias{getLabel.GenericDataFile}
\alias{getLastAccessedOn}
\alias{getLastModifiedOn}
\alias{getListOfFullNameTranslators}
\alias{getListOfFullNameTranslators.FullNameInterface}
\alias{getNames}
\alias{getNames.GenericDataFileSetList}
\alias{getOutputExtension}
\alias{getOutputExtension.GenericDataFile}
\alias{getPathname}
\alias{getPathnames}
\alias{getReadArguments}
\alias{getReadArguments.TabularTextFile}
\alias{getSet}
\alias{getSet.GenericDataFileSetList}
\alias{getSets}
\alias{getSets.GenericDataFileSetList}
\alias{getSubdirs}
\alias{getSubdirs.GenericDataFileSet}
\alias{getTags}
\alias{getTags.GenericDataFileSetList}
\alias{hasBeenModified}
\alias{hasColumnHeader}
\alias{hasFile}
\alias{hasFile.GenericDataFileSet}
\alias{hasTag}
\alias{hasTags}
\alias{indexOf}
\alias{indexOf.GenericDataFileSetList}
\alias{length.GenericDataFileSetList}
\alias{name}
\alias{name.default}
\alias{nbrOfColumns}
\alias{nbrOfFiles}
\alias{nbrOfFiles.GenericDataFileSetList}
\alias{nbrOfLines}
\alias{nbrOfRows}
\alias{nbrOfSets}
\alias{nbrOfSets.GenericDataFileSetList}
\alias{readChecksum}
\alias{readColumnNames}
\alias{readColumnNames.TabularTextFile}
\alias{readColumns}
\alias{readColumns.GenericTabularFile}
\alias{readColumns.TabularTextFile}
\alias{readDataFrame}
\alias{readLines}
\alias{readLines.default}
\alias{readRawHeader}
\alias{readRawHeader.TabularTextFile}
\alias{readTableHeader}
\alias{readTableHeader.default}
\alias{renameTo}
\alias{renameToUpperCaseExt}
\alias{renameToUpperCaseExt.GenericDataFile}
\alias{reorder.GenericDataFileSet}
\alias{sapply}
\alias{sapply.GenericDataFileSet}
\alias{sapply.default}
\alias{setAlias}
\alias{setAlias.GenericDataFile}
\alias{setAttribute}
\alias{setAttribute.GenericDataFile}
\alias{setAttributes}
\alias{setAttributes.GenericDataFile}
\alias{setAttributesBy}
\alias{setAttributesBy.GenericDataFile}
\alias{setAttributesByTags}
\alias{setAttributesByTags.GenericDataFile}
\alias{setColumnNameTranslator}
\alias{setColumnNameTranslator.GenericTabularFile}
\alias{setCommentChar}
\alias{setCommentChar.TabularTextFile}
\alias{setDepth}
\alias{setDepth.GenericDataFileSet}
\alias{setExtensionPattern}
\alias{setExtensionPattern.GenericDataFile}
\alias{setFullName}
\alias{setFullNameTranslator}
\alias{setFullNameTranslator.FullNameInterface}
\alias{setFullNamesTranslator}
\alias{setFullNamesTranslator.GenericDataFileSet}
\alias{setLabel.GenericDataFile}
\alias{setListOfFullNameTranslators}
\alias{setListOfFullNameTranslators.FullNameInterface}
\alias{setName}
\alias{setTags}
\alias{setTags.GenericDataFileSetList}
\alias{sortBy}
\alias{tags}
\alias{tags.default}
\alias{testAttributes}
\alias{testAttributes.GenericDataFile}
\alias{translateColumnNames}
\alias{translateColumnNames.GenericTabularFile}
\alias{translateFullName}
\alias{translateFullName.FullNameInterface}
\alias{update2}
\alias{update2.GenericDataFileSet}
\alias{updateFullName}
\alias{updateFullName.FullNameInterface}
\alias{updateFullName.GenericDataFileSet}
\alias{updateFullNames}
\alias{updateFullNames.GenericDataFileSet}
\alias{validateChecksum}
\alias{verify}
\alias{verify.GenericTabularFile}
\alias{verify.TabularTextFile}
\alias{writeChecksum}
\alias{writeColumnsToFiles}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
